Lecture2

+++ Matchismo/Matchismo/Model/Card.m
@@ -10,9 +10,9 @@
 
 @implementation Card
 
-- (int)match:(NSArray *)otherCards
+- (NSInteger)match:(NSArray *)otherCards
 {
-    int score = 0;
+    NSInteger score = 0;
     
     for (Card *card in otherCards) {
         if ([card.contents isEqualToString:self.contents]) {
